requirements.txt
apturl==0.5.2
backcall==0.1.0
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-rest-auth==0.9.5
djangorestframework==3.9.4
feedparser==5.1.3
GDAL==1.11.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.0
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
numpy==1.11.0
oauthlib==3.0.1
onboard==1.2.0
padme==1.1.1
parso==0.4.0
pexpect==4.0.1
pickleshare==0.7.5
Pillow==3.1.2
plainbox==0.25
prompt-toolkit==2.0.9
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.4.2
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-debian==0.1.27
python-systemd==231
python3-openid==3.1.0
pytz==2019.1
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
requests-oauthlib==1.2.0
sessioninstaller==0.0.0
six==1.10.0
sqlparse==0.3.0
ssh-import-id==5.5
system-service==0.3
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
wcwidth==0.1.7
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
